==> nbs/05_agg-processor.ipynb <==
# Preprocessing 

import os
import sys
import numpy as np
import pandas as pd
from tqdm import tqdm
from multiprocessing import Pool, cpu_count

sys.path.append("../src/")

from utilities import *

pd.set_option("display.max_columns", None)

filepath = os.path.join(dir_dict["feature_engineered"], f"completed_feateng.parquet")
df = pd.read_parquet(filepath)

df["fighter_name"] = df["fighter_name"].astype("category")

df.select_dtypes(exclude=["number", "bool"])

to_drop = ["event_name", "event_city", "event_state", "event_country",
           "fight_bout", "fight_details", "fighter_dob",  
           "fighter_name", "fighter_nickname"]

df = df.drop(to_drop, axis=1)

### One Hot Encoding 

cat_cols = df.select_dtypes(exclude=["number", "bool"]).columns.to_list()
cat_cols.remove("event_date")
cat_cols

df[cat_cols] = df[cat_cols].astype("category")

df = pd.concat([df, pd.get_dummies(df[cat_cols])], axis=1)

df = df.drop(cat_cols, axis=1)

bool_cols = df.select_dtypes("bool").columns
df[bool_cols] = df[bool_cols].astype("uint8")

### Drop Fight columns

to_drop = [col for col in df.columns if "fight_" in col] 
to_drop.remove("fight_id")
to_drop.remove("fight_fighter_win")
df = df.drop(to_drop, axis=1)

### Deal with Null Values

df.isna().sum().sort_values()

null_counter = df.isna().sum()

null_pct = null_counter.sort_values() / df.shape[0]

too_much_missing = null_pct[null_pct > 0.5].index.to_list()

len(too_much_missing)

Round4 or Round5 columns

round_4_5_cols = [col for col in too_much_missing if "round4" in col or "round5" in col]
len(round_4_5_cols)

df = df.drop(round_4_5_cols, axis=1)

null_counter = df.isna().sum()

null_pct = null_counter.sort_values() / df.shape[0]

null_pct

desc_cols = ["fighter_height_inches", "fighter_age", "fighter_reach_inches"]
df = df.loc[df[desc_cols].notna().all(axis=1)].reset_index(drop=True)

null_counter = df.isna().sum()

null_pct = null_counter.sort_values() / df.shape[0]

null_cols = null_pct[null_pct > 0].index.to_list()

len(null_cols)

len([col for col in null_cols if "cummean" in col])

#### Fighter Stats Columns

if null_cols:
    stat_cols = [col for col in null_cols if "fighter" in col]

    df[stat_cols] = \
        df.groupby("fighter_id")[stat_cols].transform(lambda x: x.fillna(method="ffill")\
                                                                          .fillna(method="bfill"))

df.isna().sum().sort_values()

if null_cols:
    df[stat_cols] = df[stat_cols].fillna(df[stat_cols].mean())

df.isna().sum().sort_values()

filepath = os.path.join(dir_dict["preprocessed"], f"agg_processed.parquet")
df.to_parquet(filepath)
