==> nbs/09_feature-importances.ipynb <==
## Final Results

import os
import sys
import json
import time
import joblib
import warnings
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt
from multiprocessing import Pool, cpu_count

sys.path.append("../src/")

from utilities import *

pd.set_option("display.max_columns", None)

from sklearn.model_selection import cross_val_score

### Load Data

filepath = os.path.join(dir_dict["train_test"], f"train.parquet")
train = pd.read_parquet(filepath)
filepath = os.path.join(dir_dict["train_test"], f"test.parquet")
test = pd.read_parquet(filepath)

df = pd.concat([train, test], axis=0)
to_drop = ["event_date", "fight_id", "fighter_id", "opponent_id"]
df = df.drop(to_drop, axis=1)

target = "fight_fighter_win"
X, y = df.drop(target, axis=1), df[target]

### Load Pipelines

pipe = joblib.load("../assets/model_training/trained_pipeline_final.joblib")

## Feature Importances

selector = pipe.named_steps["selector"]

_ = selector.fit_transform(X, y)

final_features = selector.get_feature_names_out()

model = pipe.named_steps["model"]

coefs = model.coef_[0]

feature_coefs = pd.Series(coefs, index=final_features)

simple_features = [col for col in feature_coefs.index if "cummean" not in col]
feature_coefs = feature_coefs[simple_features]

pos_features = feature_coefs.sort_values(ascending=False)[:10]
empty_row = pd.Series(["..."], index=["..."])
neg_features = feature_coefs.sort_values()[:10].sort_values(ascending=False)
imp_features = pd.DataFrame(pd.concat([pos_features, empty_row, neg_features]), columns=["feature_coef"])

imp_features.index.name = "feature"

imp_features

#### Younger Fighter

younger_fighter_winloss = df.loc[df["fighter_age"] < df["opponent_age"], target]

younger_fighter_winloss.sum() / younger_fighter_winloss.shape[0]

#### Reach

reach_adv_fighter_winloss = df.loc[df["fighter_reach_inches"] > df["opponent_reach_inches"], target]

reach_adv_fighter_winloss.sum() / reach_adv_fighter_winloss.shape[0]

corrs = df.corr()[target]

corrs.sort_values(ascending=False)["fighter_win_streak"]
