==> nbs/10_inference.ipynb <==
import os
import sys
import json
import time
import joblib
import warnings
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt
from multiprocessing import Pool, cpu_count

sys.path.append("../src/")

from utilities import *

pd.set_option("display.max_columns", None)

### Load Pipelines

pipe = joblib.load("../assets/model_training/trained_pipeline.joblib")

### Load Data

filepath = os.path.join(dir_dict["preprocessed"], f"upcoming_processed.parquet")
df = pd.read_parquet(filepath)
print(df.shape)

to_drop = ["event_date", "fight_id", "fighter_id", "opponent_id"]
identifiers = df[to_drop]
df = df.drop(to_drop, axis=1)

y_pred = pipe.predict(df)

df["predicted_win"] = y_pred
df = pd.concat([identifiers, df], axis=1)

df = df.drop_duplicates("fight_id", ignore_index=True)
df = df.sort_values("event_date")

filepath = os.path.join(dir_dict["clean"], f"fighters.parquet")
fighters = pd.read_parquet(filepath)

df = df.merge(fighters[["fighter_id", "fighter_name"]])
df = df.merge(fighters[["fighter_id", "fighter_name"]].rename(lambda s: s.replace("fighter", "opponent"), axis=1))

df[["fighter_name", "opponent_name", "predicted_win"]]
